+++ lib/src/features/entries/application/entries_service.dart
@@ -2,13 +2,13 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/authentication/data/firebase_auth_repository.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/authentication/domain/app_user.dart';
-import 'package:starter_architecture_flutter_firebase/src/features/entries/model/daily_jobs_details.dart';
-import 'package:starter_architecture_flutter_firebase/src/features/entries/model/entries_list_tile_model.dart';
-import 'package:starter_architecture_flutter_firebase/src/features/entries/model/entry_job.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/entries/domain/daily_jobs_details.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/entries/domain/entries_list_tile_model.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/entries/domain/entry_job.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/jobs/data/firestore_repository.dart';
 import 'package:starter_architecture_flutter_firebase/src/utils/format.dart';
-import 'package:starter_architecture_flutter_firebase/src/features/jobs/models/entry.dart';
-import 'package:starter_architecture_flutter_firebase/src/features/jobs/models/job.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/jobs/domain/entry.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/jobs/domain/job.dart';
 
 // TODO: Clean up this code a bit more
 class EntriesService {
