+++ lib/src/features/entries/domain/entries_list_tile_model.dart
