+++ lib/src/features/entries/domain/entry_job.dart
@@ -1,0 +1,9 @@
+import 'package:starter_architecture_flutter_firebase/src/features/jobs/domain/entry.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/jobs/domain/job.dart';
+
+class EntryJob {
+  EntryJob(this.entry, this.job);
+
+  final Entry entry;
+  final Job job;
+}
