--- lib/src/features/entries/model/entry_job.dart
@@ -1,9 +1,0 @@
-import 'package:starter_architecture_flutter_firebase/src/features/jobs/models/entry.dart';
-import 'package:starter_architecture_flutter_firebase/src/features/jobs/models/job.dart';
-
-class EntryJob {
-  EntryJob(this.entry, this.job);
-
-  final Entry entry;
-  final Job job;
-}
