+++ lib/src/features/jobs/domain/entry.dart
