+++ lib/src/features/jobs/domain/job.dart
